requirements.txt
# Core dependencies
pydantic>=2.11.5
pillow
python-dotenv>=1.0.0
httpx>=0.24.0
rich

# Browser automation
playwright
webdriver-manager
pyautogui

# AI/ML dependencies
openai>=1.0.0

# Finetuned UI-TARS model dependencies
torch>=2.0.0
transformers>=4.40.0
peft>=0.7.0
accelerate>=0.20.0
safetensors>=0.3.0

# Additional dependencies for image processing
numpy
requests
